coinance_backend/pyproject.toml
[tool.poetry]
name = "coinance_backend"
version = "0.1.0"
description = ""
authors = ["Dongho Yu <n0rr7882@gmail.com>"]

[tool.poetry.dependencies]
python = "3.8"
django-filter = "^2.2.0"
djangorestframework = "^3.11.0"
markdown = "^3.2.1"
djangorestframework-simplejwt = "^4.4.0"
django = "^2.2.10"
python-dotenv = "^0.11.0"
mysqlclient = "^1.4.6"
celery = {extras = ["redis"], version = "^4.4.0"}
django-celery-results = "^1.2.0"
channels = "^2.4.0"
django-celery-beat = "^1.6.0"
requests = "^2.22.0"
django-model-utils = "^4.0.0"
websocket-client = "^0.57.0"
channels-redis = "^2.4.2"
django-redis = "^4.11.0"
rest-social-auth = "^4.1.0"

[tool.poetry.dev-dependencies]
django-debug-toolbar = "^2.2"
pylint = "^2.4.4"
autopep8 = "^1.5"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
